refactor: Amend naming of training related functionalities
Change `training-sessions` to `trainings` for the model, controller, table and route and type definition. Remove duration attribute and set venue of training based on mode.
